Move Data.dataload() to utils module
